+++ server/src/main/java/io/crate/statistics/TableStatsService.java
@@ -21,86 +21,132 @@
 
 package io.crate.statistics;
 
-import static io.crate.common.collections.Iterables.sequentialStream;
 
-import io.crate.metadata.RelationName;
-import io.crate.metadata.table.TableInfo;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.elasticsearch.Version;
+import org.elasticsearch.cluster.service.ClusterService;
+import org.elasticsearch.common.inject.Inject;
+import org.elasticsearch.common.inject.Singleton;
+import org.elasticsearch.common.settings.Setting;
+import org.elasticsearch.common.settings.Setting.Property;
+import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.unit.ByteSizeUnit;
+import org.elasticsearch.common.unit.ByteSizeValue;
+import org.elasticsearch.threadpool.Scheduler;
+import org.elasticsearch.threadpool.ThreadPool;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.VisibleForTesting;
 
-import java.util.Map;
-import java.util.stream.Stream;
+import io.crate.session.BaseResultReceiver;
+import io.crate.session.Session;
+import io.crate.session.Sessions;
+import io.crate.common.unit.TimeValue;
+import io.crate.data.Row;
 
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
+/**
+ * Periodically refresh {@link CachedTableStatsService} based on {@link #refreshInterval}.
+ */
+@Singleton
+public class TableStatsService implements Runnable {
 
+    private static final Logger LOGGER = LogManager.getLogger(TableStatsService.class);
 
-public class TableStatsService implements TableStats {
+    public static final Setting<TimeValue> STATS_SERVICE_REFRESH_INTERVAL_SETTING = Setting.timeSetting(
+        "stats.service.interval", TimeValue.timeValueHours(24), Property.NodeScope, Property.Dynamic, Property.Exposed);
 
-    private final TableStatsPersistenceService tableStatsPersistenceService;
+    public static final Setting<ByteSizeValue> STATS_SERVICE_THROTTLING_SETTING = Setting.byteSizeSetting(
+        "stats.service.max_bytes_per_sec", new ByteSizeValue(40, ByteSizeUnit.MB), Property.NodeScope, Property.Dynamic, Property.Exposed);
 
-    private final Cache<RelationName, Stats> cache = Caffeine.newBuilder()
-        .maximumSize(100)
-        .build();
+    static final String STMT = "ANALYZE";
 
-    public TableStatsService(TableStatsPersistenceService tableStatsProvider) {
-        this.tableStatsPersistenceService = tableStatsProvider;
+    private final ClusterService clusterService;
+    private final ThreadPool threadPool;
+    private final Sessions sessions;
+
+    private Session session;
+
+    @VisibleForTesting
+    volatile TimeValue refreshInterval;
+
+    @VisibleForTesting
+    volatile Scheduler.ScheduledCancellable scheduledRefresh;
+
+    @Inject
+    public TableStatsService(Settings settings,
+                             ThreadPool threadPool,
+                             ClusterService clusterService,
+                             Sessions sessions) {
+        this.threadPool = threadPool;
+        this.clusterService = clusterService;
+        this.sessions = sessions;
+        refreshInterval = STATS_SERVICE_REFRESH_INTERVAL_SETTING.get(settings);
+        scheduledRefresh = scheduleNextRefresh(refreshInterval);
+
+        clusterService.getClusterSettings().addSettingsUpdateConsumer(
+            STATS_SERVICE_REFRESH_INTERVAL_SETTING, this::setRefreshInterval);
     }
 
-    public void updateTableStats(Map<RelationName, Stats> tableStats) {
-        tableStatsPersistenceService.deleteTableStats();
-        this.tableStatsPersistenceService.write(tableStats);
-        cache.invalidateAll();
+    @Override
+    public void run() {
+        updateStats();
     }
 
-    public void reset() {
-        tableStatsPersistenceService.deleteTableStats();
-        cache.invalidateAll();
-    }
-
-
-    public long numDocs(RelationName relationName) {
-        return fromCache(relationName).numDocs();
-
-    }
-
-    public long estimatedSizePerRow(RelationName relationName) {
-        return fromCache(relationName).averageSizePerRowInBytes();
-    }
-
-    private Stats fromCache(RelationName relationName) {
-        Stats stats = cache.get(relationName, r -> tableStatsPersistenceService.apply(relationName));
-        if (stats == null) {
-            stats = Stats.EMPTY;
+    public void updateStats() {
+        if (clusterService.localNode() == null) {
+            /*
+              During a long startup (e.g. during an upgrade process) the localNode() may be null
+              and this would lead to NullPointerException in the TransportExecutor.
+             */
+            LOGGER.debug("Could not retrieve table stats. localNode is not fully available yet.");
+            return;
         }
-        return stats;
-    }
-
-    public long estimatedSizePerRow(TableInfo tableInfo) {
-        Stats stats = fromCache(tableInfo.ident());
-        if (stats.isEmpty()) {
-            // if stats are not available we fall back to estimate the size based on
-            // column types. Therefore we need to get the column information.
-            return stats.estimateSizeForColumns(tableInfo);
-        } else {
-            return stats.averageSizePerRowInBytes();
+        if (clusterService.state().nodes().getMinNodeVersion().before(Version.V_5_7_0)) {
+            // Streaming format changed in 5.7.0, so wait until everything is upgraded before
+            // collecting stats
+            LOGGER.debug("Could not retrieve table stats.  Cluster not fully updated yet");
+            return;
         }
-    }
-
-    public Iterable<ColumnStatsEntry> statsEntries(Iterable<RelationName> relationNames) {
-        return () -> sequentialStream(relationNames)
-            .flatMap(relationName -> {
-                Stats stats = fromCache(relationName);
-                if (stats.isEmpty()) {
-                    return Stream.empty();
-                } else {
-                    return stats.statsByColumn().entrySet().stream()
-                        .map(columnEntry ->
-                            new ColumnStatsEntry(relationName, columnEntry.getKey(), columnEntry.getValue()));
+        if (!clusterService.state().nodes().isLocalNodeElectedMaster()) {
+            // `ANALYZE` will publish the new stats to all nodes, so we need only a single node running it.
+            return;
+        }
+        try {
+            BaseResultReceiver resultReceiver = new BaseResultReceiver();
+            resultReceiver.completionFuture().whenComplete((res, err) -> {
+                scheduledRefresh = scheduleNextRefresh(refreshInterval);
+                if (err != null) {
+                    LOGGER.error("Error running periodic " + STMT + "", err);
                 }
-            }).iterator();
+            });
+            if (session == null) {
+                session = sessions.newSystemSession();
+            }
+            session.quickExec(STMT, resultReceiver, Row.EMPTY);
+        } catch (Throwable t) {
+            LOGGER.error("error retrieving table stats", t);
+        }
     }
 
-    public Stats getStats(RelationName relationName) {
-        return fromCache(relationName);
+    @Nullable
+    private Scheduler.ScheduledCancellable scheduleNextRefresh(TimeValue refreshInterval) {
+        if (refreshInterval.millis() > 0) {
+            return threadPool.schedule(
+                this,
+                refreshInterval,
+                ThreadPool.Names.REFRESH
+            );
+        }
+        return null;
     }
 
+    private void setRefreshInterval(TimeValue newRefreshInterval) {
+        if (scheduledRefresh != null) {
+            scheduledRefresh.cancel();
+            scheduledRefresh = null;
+        }
+        refreshInterval = newRefreshInterval;
+        scheduledRefresh = scheduleNextRefresh(newRefreshInterval);
+    }
 }
+
