+++ server/src/test/java/io/crate/integrationtests/JoinIntegrationTest.java
@@ -45,7 +45,7 @@ import io.crate.execution.engine.join.RamBlockSizeCalculator;
 import io.crate.execution.engine.sort.OrderingByPosition;
 import io.crate.metadata.RelationName;
 import io.crate.statistics.Stats;
-import io.crate.statistics.TableStatsService;
+import io.crate.statistics.CachedTableStatsService;
 import io.crate.testing.Asserts;
 import io.crate.testing.UseHashJoins;
 import io.crate.testing.UseRandomizedOptimizerRules;
@@ -917,8 +917,8 @@ public class JoinIntegrationTest extends IntegTestCase {
         execute("insert into t2 (x) values (1), (3), (4), (4), (5), (6)");
         execute("refresh table t1, t2");
 
-        Iterable<TableStatsService> tableStatsOnAllNodes = cluster().getInstances(TableStatsService.class);
-        for (TableStatsService tableStats : tableStatsOnAllNodes) {
+        Iterable<CachedTableStatsService> tableStatsOnAllNodes = cluster().getInstances(CachedTableStatsService.class);
+        for (CachedTableStatsService tableStats : tableStatsOnAllNodes) {
             Map<RelationName, Stats> newStats = new HashMap<>();
             newStats.put(new RelationName(sqlExecutor.getCurrentSchema(), "t1"), new Stats(4L, 16L, Map.of()));
             newStats.put(new RelationName(sqlExecutor.getCurrentSchema(), "t2"), new Stats(6L, 24L, Map.of()));
@@ -946,8 +946,8 @@ public class JoinIntegrationTest extends IntegTestCase {
         execute("insert into t3 (y) values (0), (1), (2), (3), (4), (5), (6), (7), (8), (9)");
         execute("refresh table t1, t2, t3");
 
-        Iterable<TableStatsService> tableStatsOnAllNodes = cluster().getInstances(TableStatsService.class);
-        for (TableStatsService tableStats : tableStatsOnAllNodes) {
+        Iterable<CachedTableStatsService> tableStatsOnAllNodes = cluster().getInstances(CachedTableStatsService.class);
+        for (CachedTableStatsService tableStats : tableStatsOnAllNodes) {
             Map<RelationName, Stats> newStats = new HashMap<>();
             newStats.put(new RelationName(sqlExecutor.getCurrentSchema(), "t1"), new Stats(2L, 8L, Map.of()));
             newStats.put(new RelationName(sqlExecutor.getCurrentSchema(), "t2"), new Stats(3L, 12L, Map.of()));
@@ -997,7 +997,7 @@ public class JoinIntegrationTest extends IntegTestCase {
         long tableSizeInBytes = new Random().nextInt(3 * (int) availableMemory);
         long rowSizeBytes = tableSizeInBytes / rowsCount;
 
-        for (TableStatsService tableStats : cluster().getInstances(TableStatsService.class)) {
+        for (CachedTableStatsService tableStats : cluster().getInstances(CachedTableStatsService.class)) {
             Map<RelationName, Stats> newStats = new HashMap<>();
             newStats.put(new RelationName(sqlExecutor.getCurrentSchema(), relationName), new Stats(rowsCount, tableSizeInBytes, Map.of()));
             tableStats.updateTableStats(newStats);
