+++ server/src/testFixtures/java/io/crate/testing/TestingHelpers.java
@@ -30,7 +30,6 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.lang.reflect.Array;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -60,7 +59,6 @@ import io.crate.analyze.TableParameters;
 import io.crate.common.collections.Lists;
 import io.crate.data.Row;
 import io.crate.execution.ddl.tables.MappingUtil;
-import io.crate.lucene.CrateLuceneTestCase;
 import io.crate.metadata.ColumnIdent;
 import io.crate.metadata.DocReferences;
 import io.crate.metadata.Functions;
@@ -75,8 +73,6 @@ import io.crate.metadata.settings.session.SessionSettingRegistry;
 import io.crate.planner.optimizer.LoadedRules;
 import io.crate.role.Role;
 import io.crate.statistics.StubTableStats;
-import io.crate.statistics.TableStatsService;
-import io.crate.statistics.TableStatsPersistenceService;
 import io.crate.types.DataType;
 import io.crate.types.DataTypes;
